% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_function.R
\name{as_function}
\alias{as_function}
\title{Convert to function}
\usage{
as_function(x, env = caller_env())
}
\arguments{
\item{x}{A function or formula, see [rlang::as_function()] for more
information}

\item{env}{Environment in which to fetch the function in case `x` is a string}
}
\description{
This function is a wrapper around [rlang::as_function()] which adds
a few extra features

* formulas can use `.t` in place of `.x` to be easier to understand
 in time-based functions
* functions can take additional named arguments.
}
\examples{

f1 <- as_function(~.t + 1)
f1(10)

f2 <- as_function(~.t + x)

f2(10,x=2)

}
