% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blueprint.R
\name{blueprint}
\alias{blueprint}
\alias{is_blueprint}
\title{Blueprint a Dataset}
\usage{
blueprint(...)

is_blueprint(bp)
}
\arguments{
\item{...}{arguments used to generate the blueprint}

\item{bp}{Obj to check}
}
\value{
a function which can produce a tibble based on the provided
blueprint
}
\description{
Allows for the generation of population based on a prescribed
}
\examples{
make_tbl <- blueprint(
  x = r_norm(),
  y = r_norm()
)

make_tbl(n = 2)

make_tbl(n = 5)

make_tbl2 <- blueprint(
  x = r_norm(mean = x_mu),
  y = r_unif(min = y_min, max = y_max)
)

make_tbl2(x_mu = 10, y_min = -10, y_max = -5)
is_blueprint(make_tbl)
}
