% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_n.R
\name{default_n}
\alias{default_n}
\alias{tibble_n}
\alias{dplyr_n}
\alias{args_n}
\title{Find the Default value for n in Context}
\usage{
default_n(...)

tibble_n()

dplyr_n()

args_n(...)
}
\arguments{
\item{...}{parameters to check the length of}
}
\description{
Checks for various information surrounding the call to this function
to figure out what value for n should be used
}
\section{Functions}{
\itemize{
\item \code{default_n}: Checks for the default value of n in a
context aware manner.

\item \code{tibble_n}: If we are inside of a call to  \code{tibble()},
it returns the number of rows in the \code{tibble}, otherwise
\code{NULL}

\item \code{dplyr_n}: If we are in \code{dplyr} verb function, it
returns the number of rows in the current data, otherwise \code{NULL}

\item \code{args_n}: Checks the lengths of the non-\code{NULL}
arguments passed to it. Errors if they are inconsistent.
}}

\examples{
default_n()
tibble::tibble(id = 1:3, n = default_n())
df <- tibble::tibble(id = 1:4)
dplyr::mutate(df, n = default_n())
default_n(1:5)
set_n(10)
default_n()
tibble::tibble(id = 1:3, n = tibble_n())
dplyr::mutate(df, n = dplyr_n())
args_n(1:3)
\dontrun{
args_n(1:3, 1:4)
}

}
