% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_dots.R
\name{extract_dots}
\alias{extract_dots}
\title{Extract the ellipsis inside a function}
\usage{
extract_dots()
}
\description{
Allow the named entries in `...` to be used easily within a
function by attaching them to the function's environment
}
\examples{

\dontrun{
f <- function(...) {
  a + b
}

# Throws an error because a and b are trapped inside ...
f(a = 1, b = 2)
}

f <- function(...) {
  extract_dots()
  a + b
}
f(a = 1, b = 2)
}
