% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit.R
\name{logit}
\alias{logit}
\alias{invlogit}
\title{The logit and inverse logit functions}
\usage{
logit(prob, base = exp(1))

invlogit(alpha, base = exp(1))
}
\arguments{
\item{prob}{vector of probabilities}

\item{base}{base of the logarithmic function to use}

\item{alpha}{vector of values to find the inverse logit of}
}
\value{
A numeric vector
}
\description{
Calculates the logit or the inverse logit of a value
}
\examples{

logit(0.5)

logit(seq(0.01, 0.99, 0.01))
invlogit(-10:10)
}
