% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/built_in_distributions.R
\name{r_norm}
\alias{r_norm}
\title{Generate Normally Distributed Values}
\usage{
r_norm(mean = 0, sd = 1, ..., n = default_n(mean, sd), .seed = NULL)
}
\arguments{
\item{mean}{vector of means}

\item{sd}{vector of standard deviations, strictly positive}

\item{...}{Unused}

\item{n}{number of observations to generate. The [default_n()] function will
provide a default value within context}

\item{.seed}{One of the following:

* NULL (default) will not change the current seed. This is the
usual case for generating random numbers.

* A numeric value. This will be used to set the seed before generating
the random numbers. This seed will be stored with the results.

* TRUE. A random seed value will be generated and set as the seed
 before the results are generated. Again, this will be stored with
 the results.

To extract the random seed from a previously generated set of
values, use `pull_seed()`}
}
\description{
Generates a set of Normally distributed values.
}
\examples{

set_n(5)

r_norm(10)

r_norm(10,2)

r_norm(1:10)

r_norm(-2,n=10)


}
