% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rando-package.R
\name{rando-package}
\alias{rando-package}
\title{Context Aware Random Number Generation}
\description{
rando is designed to make random number generation easier
by providing the ability to set a default number of numbers to
generate or to assess the context in which the functions are being
ran.
}
