% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_n.R
\name{set_n}
\alias{set_n}
\alias{default_n}
\title{Set the default value for n}
\usage{
set_n(n)

default_n(...)
}
\arguments{
\item{n}{value to set as the default n for random number generating}

\item{...}{parameters passed from random number generating functions}
}
\description{
This function sets the default value for n, the number of random
numbers to generate. If n is not specified nor implied by context,
this is the default value that will be used.
}
\details{
The \code{default_n()} function looks for a possible value for
\code{n} in multiple places, in order, these locations are:

1. Assesses whether it is being called in a \code{tibble()} call and
   calculate how many rows the are being generated.

2. If this fails, it checks for a value of \code{dplyr::n()},
   which returns the number of rows of the current data structure
   if we are inside a \code{dplyr} verb, such as [dplyr::mutate()]

3. If parameters supplied to \code{default_n()}, then the length of
   these arguments will be used to define \code{n}

4. The last check is to see if \code{set_n()} has defined a default
   \code{n} to be used globally.
}
\examples{
set_n(100)

length(r_norm())

set_n(10)
length(r_norm())

default_n()


}
