% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_n.R
\name{set_n}
\alias{set_n}
\alias{get_n}
\title{Set the default value for n}
\usage{
set_n(n)

get_n()
}
\arguments{
\item{n}{value to set as the default n for random number generating}
}
\description{
Set and get the global value for n
}
\examples{
set_n(100)
get_n()
}
