% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_n.R
\name{set_n}
\alias{set_n}
\alias{get_n}
\title{Set and Get the Default Value for n}
\usage{
set_n(n)

get_n()
}
\arguments{
\item{n}{value to set as the default n}
}
\value{
The current \emph{global} default value for n.\cr
\code{set_n()} returns this value invisibly
}
\description{
Set and get the global value for n for rando functions
}
\examples{
set_n(100)

get_n()

}
